sysutils/boot-delay: this is neat...
So now that we have a non-persistent boot delay in our install
images we may still have users that can't live without it so
this plugin installs a persistent boot delay.

The first boot will work, can bring up network and install!

Yes, now you're saying this is stupid, which it it if you're
not seeing the bigger picture.  This paves the way for plugin-
based optimisation for the following use case:

https://github.com/opnsense/plugins/issues/9

Yes, yes, more work needs to be done but it's a start.  :)