automate the template regeneration for provided templates for #46
Unfortunately, we need to reload all subdirs (minus the vendor
directory).  The "*" matcher is not very clever and cannot be
coerced to match sub directories without potentially matching any
directories longer than their name, e.g. "Sample*" vs "Sample2".

Problem is when these are installed and the templates are wrongly
regenerated.  In theory this doesn't have any bad side effects,
but it may be a potential pitfall to avoid now rather than later.