bugfix/247 Improve Zerotier Start/Stop Behaviour
https://github.com/opnsense/plugins/issues/247

This commit improves upon the way that Zerotier starts and stops and how
networks are added/removed and activated/deactivated. There is now a "Global"
tab on the Zerotier page that enables or disables the service which is
honoured between reboots. Additionally, the service needs to be active in
order for networks to be added and removed. If the service is not active, then
the tab "Networks" will be disabled and the user will not be able to
add/remove or activate/deactivate Zerotier networks.

This should fix the observed problems raised in the issue.

The "Global" tab will be extended later to include further information and
actions, especially when Zerotier API usage is developed.

-=david=-

(cherry picked from commit 3fdca054d09bbabac124bb7ce0fc3bf533fcdaa3)
(cherry picked from commit e933dfb8af3caf52024c23e1fffc12e40724fb39)