net/zerotier: Add Basic API Functionality
This commit brings basic API functionality to the Zerotier plugin. It allows
the user to retrieve "Global" Zerotier information, peers that are talking
with this node and details on configured (and enabled) networks.

This version of the plugin provides a space for the Zerotier API key to be
input and saved, but it is not used for Basic API functionality. The
zerotier-cli client is used directly to query for global, peers and network
information.

If the API functionality is extended to allow for creating and modifying
Zerotier information from this plugin, then the API key will be required at
that point. It is introduced here for the sake of convenience for later down
the line.

Various other small changes include general code cleanups, renaming of some UI
facing labels and improved layout design.

-=david=-

Fixes: #260