Updates to the scp-backup plugin.
A private key is now stored along with the config that is the other side of
the public/private keypair for the remote host. The public key must be added
to the remote authorized_keys file.

It now also optionally will perform a StrictHostKeyCheck for connections to
the remote host. If StrictHostKeyCheck is enabled, then a HostKey needs to be
provided which is used to verify the remote host (same format as the entry in
the known_hosts file).

Both the private key and the host key are stored as entries in the config and
wrote out as tempoary files to initiate the connection. The temporary files
are removed after the backup has been performed.

-=david=-