Rebellion theme update
This commit adds the external CSS files such as nv.d3.css to the rebellion css folder. With a commit in the core I am about to do this will fix all external references EXCEPT the ones in the *gateways*.php files.

There are also a couple of minor color changes that make items more legible.

Even without the core commit this update should be pulled as the added css files will have no effect without the core pull, but they will be needed when that happens.