More informative cloudflare hostname error message
I was confused for a long time about the error message "check your hostname is
correct" because I wasn't actually setting the hostname, it was just using the
hostname of the router. It turns out that this error was raised because the
Dynamic DNS code can only update existing records, not create new ones. I think
it would be better if it could create A records but for now this more
informative error message gives another valid reason that the update could be
failing.