dns/dyndns: Fix for using apex domains with cloudflare DDNS
When refactoring the Cloudflare DDNS code, I removed a bunch of assumptions
but inadvertently added one. The refactored code did not allow for "apex"
domain records, records for which there is no hostname, just the bare
domain. (Traditionally, these have been referred to as '@' records but
with the Cloudflare API, you just hand it the name of the zone.)

Ideally there would be a checkbox or something in the UI for this, or
even better a way to specify the record and zone separately but in lieu of
that, we first try to break the provided string down into hostname and
domain portions and look up the zoneId from that. If that fails, we try
to look up the zondId using the whole string. And then finally fail if
neither return a zone from the API.

Hopefully fixes #926 and possibly the issue reported in PR #861.

(cherry picked from commit 2b4f2c8750e3fb7bf2648f90ad228dc4e087c120)
(cherry picked from commit ddb72d981d34cdd2e624e594312229f14be2dd6b)