AcmeClient: Enhanced sftp upload with configurable filenames.
Filename templates for cert, key and ca-chain can be configured in the UI and user input is sanitized to ensure valid file names are produced.
New and some existing fields are using "advanced" markers to reduce amount of fields visible in the default UI.
Also fixed minor bugs including one with unicode characters in names.