fixes SMART Status "unknown" in the smart widget and more... (#1429)
* fixes SMART Status "unknown" in the smart widget

This fixes the "Unable to detect device typ" of smartctl, because it needs the full path to the device (e.g. /dev/ada0). As second it removes white space from start of $state variable before return, which let it match in the right way in the smart_status.widget.php. And as last the additional diskinfo command let the script skip devices, where disking exits with the error 'failed, probably not a disk.'.

* replace device listing

replace fixed device pattern with list of disk devices detected by kernel