Change settings.xml to support minor debug level change
Added a note to log level help to explain the new "debug dnsapi" level. This triggers the certhelper.php to call acme.sh with "--debug 2." This additional parameter to --debug triggers acme.sh to log messages from the called DNSAPI script to report _debug messages in the script. This get picked up in the /var/log/acme.sh.log file to troubleshoot problems with interface the DNS-01 authentication script.