Clarify purpose of ID httpserver.ca
The existing description of "CA Certificate" is confusing, in that it implies the user should select the Certificate Authority responsible for generating their SSL pem (e.g. LetsEncrypt), when the parameter is solely used to specify the CA used for verifying Client Certificates. While this might be acceptable in self-signed scenarios where users have the appropriate keys to generate a client CA, it's unfortunately not the case for production SSL certs.