Interfaces_Plus - First Commit
A replacement for the standard interfaces widget. This widget adds a toggle on/off for the interface speed and interface address netmask. it will dynamically adjust the width of the columns depending on the selection. This allows for wider interface names and showing the IPv6 address full length when on a triple column display.

Added link through inc

Additions to allow link to intefaces.php

Update pkg description

Fix for wrong folder

inc files was in the wrong folder. Corrected