FRR: work in progress event handler for https://github.com/opnsense/plugins/issues/2091
Initial boilerplate code to extract interface address (vhid) status and control vtysh for status and configuration.
The code in carp_event_handler is only intended to test some callouts, the general idea is to hook InterfaceStatus() and VtySH() objects into an event handler object inherited from baseEventHandler(),
which is responsible for detecting the changes and reconfiguring the daemon.