Show handshake timestamp in system timezone (#2204)
This change shows each latest handshake time in the widget in the timezone of the system, rather than UTC as currently. The change is intended to implement this forum request: https://forum.opnsense.org/index.php?topic=20530.msg95463#msg95463.

As a disclaimer, I am by no means a PHP programmer, but this commit is based on my research. It works when I tested it.