dns/dyndns: fix copy and paste unprotected object access
The original code was intended to look up results or let them
go in case of invalid returns but it was simplified to the point
of not being able to check for validity in the return data.

Fix this the best way we can for all three affected services.

PR: https://github.com/opnsense/plugins/pull/2288
PR: https://github.com/opnsense/plugins/issues/1564