Fix issue with Godaddy Ipv6 not working
After much hunting around, the variable  $this->_dnsRequestIfIP is used in one place only, and that is  curl_setopt($ch, CURLOPT_INTERFACE, $this->_dnsRequestIfIP);

CURLOPT_INTERFACE specifies an interface, not an IP address. For some reason, this works for godaddy V4 updates but not for IPv6. Changing it to setting the interface corrects the issue. TBH, I'm baffled why only a couple of us have noticed the problem, perhaps we are the only ones using IPv6?