security/acme-client: fix calculation of renewal date, closes #2721
Now we read the validFrom information directly from the cert file
in order to calculate the renewal date.

This is necessary, because in ae697392293e4a7fb3e9ed0450a559adccbab2e6
we made the import feature available to the end-user. As a result,
the value of lastUpdate() does not only change after issue/renewal,
but also everytime the user clicks on the "import" button.