Prefix hostname to config file before committing
This MR prefixes the hostname to the config.xml file commited to git.

For users with more than one OPNSense, all utilising git backup, it is
not neccessarily practical to have one repository or one branch per OPNSense
instance. This MR prefixes the hostname to the config file to enable
more than one config in each repository