dns/ddclient: Fix IPv6 address parsing from external service (#3217)
When using external services to determine the public IPv6 address of an
interface, the bounds are too limited for short (but valid) IPv6
addresses. Reducing the necessary amount of colon characters in the
resolved IPv6 address for the `checkip` script to accept them fixes
this issue.

(cherry picked from commit 134097983f6841938e500e747f484abcb0e91921)