dns-overview.css color theme fixes for Cicada/Vicuna (#3295)
* dns-overview.css color theme fixes for Cicada/Vicuna

by the way......for the dns-overview html basic structure the color "white" was written hardcoded. the color is defined differently in each theme and is already defined in the main.css. I would ask you to remove this hardcoded part so that the colors of the themes can be used again. the part I mentioned is here. see the screenshot