Cloudflare implementation for OPNsense backend ddclient
This is just a basic implementation of cloudflare for ddclient with OPNsense backend. It supports a single hostname/record
It needs some better error handling and there is also no support for multiple Hostnames and Wildcard at the moment.
Credentials have remained the same, as with ddclient backend (Username=Mailaddress of the cloudflare account, Password=Global API Key)