dns/bind: refine previous
It was decided we only want the 'ports' trick for DNS so make
sure the prerequistes for it match the reality of the setup.
This was we can also extend the validation of the DNS port like
we are going to do for Unbound to ensure a functional DNS setup
when multiple DNS servers are being used on the same box.