dns/dnscrypt-proxy: change core DNS approach slightly
Only trigger on 0.0.0.0/:: combination.  Also makes clashes with
previous defaults and user input less likely.

Core services are already validated as being run on 0.0.0.0/:: even
when they are not so this is fine.

Change the defaults as well as they make more sense (but keep the
non-standard port) and make it a required setting (not sure what
the default would have been).