dns/bind: prevent port overlap
Although it's possible to use e.g. port 53 with selective listening
addresses its use is highly discouraged and leads to setup woes when
the user is not well-versed in these matters.  So we lock the port
in case another service is using it and if something needs to overlap
port forward rule use is highly encouraged to make that more clear.