dns/ddclient: add missing dynurl for cloudns; closes #3430
* Remove the "default" handling for backend: it was properly migrated and
  is required.  It will make flipping to the native backend easier later.
* Make the ddclient.conf file easier to read: global block, then one block
  for each account.
* Adjust the templating to sprinke missing "\" at the end of the line and
  replace hardcoded values where we can easily use service.name.

Switching back and forth between backends is a bit annoying as the other
backend can break using the correct account data from the other one, because
not all the services are shared equally between the two.

Maybe long term it would be better to build a plugin variant for ddclient
and dyndns so we could revive os-dyndns and just share the code in this one
directory.  The backend switch would then be carried out over which plugin
is installed a bit similar to what WireGuard is doing.