www/caddy: add support for changing host headers
Since (most) headers retain their original value when being proxied, it
is often necessary to override the Host header with the configured
upstream address when proxying to HTTPS, such that the Host header
matches the TLS ServerName value. It is also useful when proxying to
virtual hosts that are served from the same instance.