www/OPNProxy - move proxy access management feature to the community version.
This plugin uses squid's external acl helpers and redis to query policies quicker and more lightweight.
It has been part of the business edition for some time, but due to recent changes, it makes sense to add it to the community version as well.

Due to the redis requirement, we keep it a separate plugin, so existing setups won't start to pull redis in unexpected.

current documentation: https://docs.opnsense.org/vendor/deciso/opnproxy.html