Update reverse_proxy.volt
Add a first version of a filter functionality by domain. In this version, only handlers are filtered by domain.

A selectpicker with multi selection can choose domains, and the filter function compares these UUIDs to the UUIDs of the "reverse" UUIDs of the model relation fields. Either all domains are shown, or only elements where the UUIDs match.