www/caddy: Selectpicker that can filter the Domain, Subdomain and Handlers by selected Domain (#3937)
* Update reverse_proxy.volt

Add a first version of a filter functionality by domain. In this version, only handlers are filtered by domain.

A selectpicker with multi selection can choose domains, and the filter function compares these UUIDs to the UUIDs of the "reverse" UUIDs of the model relation fields. Either all domains are shown, or only elements where the UUIDs match.

* Update ReverseProxyController.php

A new api endpoint for the domain search selectpicker has been created. It returns the ID and a Domain+Port combination.

The search function for the Handler now returns all fields if no filter has been set, or only the referenced UUIDs when a filter has been set.

* Update ReverseProxyController.php - Add search function to subdomains

* Update reverse_proxy.volt - Reference Search Filter in Handlers, Domains and Subdomains

* Update ReverseProxyController.php - Add the search function to domains

* Update ReverseProxyController.php

A little bit of cleanup. $add_empty was unused so it's removed.

* Update reverse_proxy.volt - Changed margin to align selectpicker with other options.

* Small margin fix to align Selectpicker with other options

* Update reverse_proxy.volt - Restrict style to the ID of only the affected selectpicker