misc/theme-Aavanced: Theme based on AdvancedTomato project by the original author (#4189)
This is a new modern theme based on the AdvancedTomato project I
used to work on years ago. The design is blueish with conservative shadows.
Love to change it, so any ideas are welcome.

Initial release working with OPNSense 24.7.1. Any issues you find I will be more
than happy to fix. But I have been using theme as "beta" for months now and I
haven't found any specific issues yet. But more people more to find!