plugins: add conflicts to #4202
In order to devise a plan to cope with plugins that deinstall other
plugins the best approach seems to be to simply tell the core that
other plugins are conflicting.  We already have that information
so we just have to pass it along.

It's a bit similar to pkg and ports where this information is available
during build but not used later during runtime.  Conflicts are based
on installing the same file, but pkg only notices this when it starts
installing and will try to tell you again, but it could have already
kown.

Needs some glue in core as well but it will be very cool.