devel/debug: meet 'qyua', the quick-kyua command ;)
With the tool we can run tests from the source tree given the tests
(or a respective prebuilt set) have been installed in the system.

It creates "_" prefixed files in order to not taint the otherwise
installed tests and set up the few things that the build would
for source tree based test files.

Underneath eventually it runs kyua-test or kyua-debug depending
on the debug command line flag -d.  -a runs all the tests in the
source tree and -l lists them.

For now only supports .sh files and both /usr/src and /usr/tests
have to be manually managed, but more to come.