net/ndproxy: Disable promiscuous mode forced by port
We do not need this since our most common setup has "uplink_interface" and "downlink_mac_address" on the same interface, ndproxy and CPE router are same device.

Promiscuous mode would be needed if the "downlink_mac_address" is on a different device, e.g. a separate CPE router. If that is the case, users can enable it via interface settings.