make: kill the plugin lint code
For clarity we split core and plugins from the beginning although
plugins only really make sense with a core in mind.  To that end,
when dealing with plugins a core repository should always be available
in order to run the lint passes on the code.  This makes maintenance
a lot easier in the future.