security/netbird: fix selectpicker and unbreak migration
The auth key may be required but not giving a default for obvious
reasons just makes it end up without a required value anyway until
user contact.

This can probably be made more robust in the future, but requires
a bit of thought on what we validate/enforce here anyway like an
"enbable" checkbox being checked requires filling this value, but
it's also not on the same page or model even making constraints
tricky.