net/freeradius: modify VLAN fallback logic #5050
While it's nice to constrain this it's also very inflexible.
With fallback disabled we can retain values in tunnel pw and
VLAN ID.  Enabling it without these options may be weird, but
not a large problem.  Template was rearranged to generate
the most logic outcome.

Some due maintenance sprinkled on top.